update and lint
