+++ test/multiverse/suites/hybrid_agent/simulated_operations.rb
@@ -2,5 +2,5 @@
 # See https://github.com/newrelic/newrelic-ruby-agent/blob/main/LICENSE for complete details.
 # frozen_string_literal: true
 
-module NewRelicTestOperations
+module SimulatedTestOperations
 end
